Harden jar-source proxy for production-quality contribution
Modularize kotlin-lsp-proxy (uri, security, cache, framing, rewrite),
add zip-slip protection and archive path allowlisting, cache size GC,
and user settings (`sourceProxy`) stripped before LS init.

Quiet-by-default extension logging, Clippy-clean proxy CI on PRs,
SECURITY/CONTRIBUTING docs, and release workflow tweaks.
