+++ .github/workflows/proxy-ci.yml
@@ -1,0 +1,37 @@
+name: Proxy CI
+
+on:
+  push:
+    paths:
+      - "proxy/**"
+      - ".github/workflows/proxy-ci.yml"
+  pull_request:
+    paths:
+      - "proxy/**"
+      - ".github/workflows/proxy-ci.yml"
+      - "src/**"
+
+jobs:
+  test-proxy:
+    name: Test kotlin-lsp-proxy
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: dtolnay/rust-toolchain@stable
+      - name: Test
+        run: cargo test --manifest-path proxy/Cargo.toml
+      - name: Clippy
+        run: cargo clippy --manifest-path proxy/Cargo.toml -- -D warnings
+      - name: Format check
+        run: cargo fmt --manifest-path proxy/Cargo.toml -- --check
+
+  build-extension:
+    name: Build WASM extension
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: dtolnay/rust-toolchain@stable
+        with:
+          targets: wasm32-wasip1
+      - name: Build extension
+        run: cargo build --release --target wasm32-wasip1
