+++ src/language_servers.rs
@@ -1,10 +1,10 @@
 mod kotlin_language_server;
 mod kotlin_lsp;
 mod proxy;
+mod settings;
 mod util;
 
 pub use kotlin_language_server::KotlinLanguageServer;
 pub use kotlin_lsp::KotlinLSP;
-pub use proxy::{
-    absolutize_in_work_dir, find_or_install_proxy, proxy_debug_enabled, proxy_disabled,
-};
+pub use proxy::{absolutize_in_work_dir, find_or_install_proxy, proxy_disabled};
+pub use settings::{merge_proxy_settings, SourceProxySettings};
